added Literal as alias for PhpLiteral
