+++ src/PhpGenerator/PhpLiteral.php
@@ -10,23 +10,6 @@ declare(strict_types=1);
 namespace Nette\PhpGenerator;
 
 
-/**
- * PHP literal value.
- */
-class PhpLiteral
+class PhpLiteral extends Literal
 {
-	/** @var string */
-	private $value;
-
-
-	public function __construct(string $value)
-	{
-		$this->value = $value;
-	}
-
-
-	public function __toString(): string
-	{
-		return $this->value;
-	}
 }
